+++ RELEASE_NOTES.md
@@ -1,4 +1,20 @@
 **2026.26.3**
 
+**Keybinds**
+
+* You can now use the LEFT ARROW and RIGHT ARROW keys to navigate through photos.
+
 **Notifications**
-* If you get an group invite you can now open the group before you accept or deny the group invite.
+
+* When you receive a group invite, you can now open the group before accepting or denying the invite.
+
+**Design**
+
+* Redesigned the user and friend item shown across multiple places.
+* Added world location information to the user item when a friend is in a world.
+* Changed the padding between user items in the People tab to match all other tabs.
+
+**Fixes**
+
+* Fixed the context menu having lower Z-ordering than preview modals, causing it to render behind them.
+* Fixed the Current Instance card being slightly bigger than the Current World card in user profiles.